Add some basic (broken) tests

+++ rtplot/api/RealTimePlot.py
@@ -11,22 +11,37 @@ from matplotlib import style
 from collections import deque
 import multiprocessing
 
-import rtplot.internal
+import rtplot.internal as Internal
 import rtplot.helpers
+from rtplot.helpers import RtplotEvent
+
 
 class RealTimePlot:
     """
     Base class for the user facing interface to create a real time plot.
     """
 
-    def __init__(self, seconds_to_show=None, linestyle='b-', *args, **kwargs):
-        # Gotta put in the internal plot class here when you subclass
-        self._internal_plot_class = None
-        self.is_started = False
-        self._seconds_to_show = seconds_to_show
-        self._refresh_rate = 10
-        self._linestyle = linestyle
+    def __init__(self, internal_plot_type, **plot_options):
+        # Check is mandatory
+        if not issubclass(internal_plot_type, Internal.RealTimePlot):
+            raise TypeError(
+                f"Given internal plot type must subclass rtplot.internal.RealTimePlot but is type {internal_plot_type}")
+
+        self._internal_plot_type = internal_plot_type
+
+        self._timeout = plot_options.setdefault("timeout", 10)
+        self._seconds_to_show = plot_options.setdefault("seconds_to_show", 10)
+        self._refresh_rate = plot_options.setdefault("refresh_rate", 10)
+        self._linestyle = plot_options.setdefault("linestyle", 'b-')
+
+        # setdefault above inserts the keys if needed so this is fully up to date
+        self.plot_options = plot_options
         self.statics = []
+        # Link statics to plot options so it's passed down to internal plot
+        # when it gets updated by add_statics
+        self.plot_options["statics"] = self.statics
+
+        self.is_started = False
 
     def __enter__(self):
         """
@@ -40,6 +55,11 @@ class RealTimePlot:
         Allow usage in 'with' statements.
         """
         self.quit()
+        if exception_type:
+            print("Plot quit unexpectedly.")
+            print(f"{exception_type} : {exception_value}")
+            print(traceback)
+        return True  # Suppress exception
 
     @property
     def seconds_to_show(self):
@@ -54,7 +74,7 @@ class RealTimePlot:
         if not self.is_started:
             self._seconds_to_show = value
         else:
-            raise Exception(
+            print(
                 "Cannot change value while plot is running. Quit plot to change values, then restart.")
 
     @property
@@ -73,7 +93,7 @@ class RealTimePlot:
         if not self.is_started:
             self._linestyle = value
         else:
-            raise Exception(
+            print(
                 "Cannot change value while plot is running. Quit plot to change values, then restart.")
 
     @property
@@ -88,34 +108,38 @@ class RealTimePlot:
         if not self.is_started:
             self._refresh_rate = value
         else:
-            raise Exception(
+            print(
                 "Cannot change value while plot is running. Quit plot to change values, then restart.")
 
     def add_static(self, static_name, **kwargs):
+        if self.is_started:
+            print(
+                "Statics added while plot is running will not appear unless plot is quit and restarted.")
         valid_statics = ["point", "circle", "rectangle", "vline", "hline"]
         if static_name not in valid_statics:
-            raise Exception(
+            print(
                 "Invalid static requested. See help on this function.")
         new_static = {}
         if static_name == "point":
-            new_static = rtplot.core.helpers.point(**kwargs)
+            new_static = rtplot.helpers.point(**kwargs)
         elif static_name == "circle":
-            new_static = rtplot.core.helpers.circle(**kwargs)
+            new_static = rtplot.helpers.circle(**kwargs)
         elif static_name == "rectangle":
-            new_static = rtplot.core.helpers.rectangle(**kwargs)
+            new_static = rtplot.helpers.rectangle(**kwargs)
         elif static_name == "vline":
-            new_static = rtplot.core.helpers.vline(**kwargs)
+            new_static = rtplot.helpers.vline(**kwargs)
         elif static_name == "hline":
-            new_static = rtplot.core.helpers.hline(**kwargs)
+            new_static = rtplot.helpers.hline(**kwargs)
         self.statics.append(new_static)
 
-    def _create_internal_plot(self, mp_queue):
+    def _create_internal_plot(self, data_queue, message_queue):
         """
         Handles instantiating the internal plot. End users should not need to call this ever, it's
         done in the start function.
         """
-        self._plot = self._internal_plot_class(
-            mp_queue, self.seconds_to_show, self.linestyle, self.statics)
+
+        self._plot = self._internal_plot_type(
+            data_queue, message_queue, **self.plot_options)
         # Convert Hz to milliseconds
         self._plot.start_animation(int(1 / self._refresh_rate * 1E3))
 
@@ -124,13 +148,20 @@ class RealTimePlot:
         Starts the live plotter!
         """
         try:
-            self._queue = multiprocessing.Queue()
-            self.is_started = True
+            self._data_queue = multiprocessing.Queue()
+            self._message_queue = multiprocessing.Queue()
             self._process = multiprocessing.Process(
-                target=self._create_internal_plot, args=(self._queue,))
+                target=self._create_internal_plot, args=(self._data_queue, self._message_queue))
+            self._process.daemon = True
             self._process.start()
+            self.is_started = True
         except RuntimeError:
             print("""If you're seeing this error, you probably forgot to wrap your code in a `if __name == "__main__"` guard.""")
+            self.quit()
+            raise SystemExit
+        except:
+            print("Error in starting plot.")
+            self.quit()
             raise SystemExit
 
     def update(self, data):
@@ -145,5 +176,12 @@ class RealTimePlot:
         """
         End the live plotter.
         """
-        self._process.terminate()
+        print("Closing plot")
+        self._message_queue.put(RtplotEvent.REQUEST_KILL)
+        time.sleep(1)
+        # Force kill if necessary
+        if self._process.is_alive():
+            self._process.terminate()
+        self._data_queue.close()
+        self._message_queue.close()
         self.is_started = False
